feat: add inactive merchants query
