+++ core/api/src/app/merchants/remove-inactive-merchants.ts
@@ -1,40 +1,11 @@
+import { listInactiveMerchants } from "./list-inactive-merchants"
 import { deleteMerchantByUsername } from "./delete-merchant-map"
 
-import { getTransactionsForAccountByWalletIds } from "@/app/accounts"
-
-import { AccountsRepository, MerchantsRepository } from "@/services/mongoose"
-
-const INACTIVE_MONTHS = 6
-
 export const removeInactiveMerchants = async (): Promise<void | ApplicationError> => {
-  const merchantsRepo = MerchantsRepository()
-  const accountsRepo = AccountsRepository()
+  const inactiveMerchants = await listInactiveMerchants()
+  if (inactiveMerchants instanceof Error) return inactiveMerchants
 
-  const merchants = await merchantsRepo.listForMap()
-  if (merchants instanceof Error) return merchants
-
-  const sixMonthsAgo = new Date()
-  sixMonthsAgo.setMonth(sixMonthsAgo.getMonth() - INACTIVE_MONTHS)
-
-  for (const merchant of merchants) {
-    if (merchant.createdAt > sixMonthsAgo) continue
-
-    const account = await accountsRepo.findByUsername(merchant.username)
-    if (account instanceof Error) continue
-
-    const transactionsResult = await getTransactionsForAccountByWalletIds({
-      account,
-      rawPaginationArgs: { first: 1 },
-    })
-    if (transactionsResult instanceof Error) continue
-
-    const transactions = transactionsResult.edges
-
-    const lastTransactionDate =
-      transactions.length > 0 ? transactions[0].node.createdAt : new Date(0)
-
-    if (lastTransactionDate < sixMonthsAgo) {
-      await deleteMerchantByUsername({ username: merchant.username })
-    }
+  for (const merchant of inactiveMerchants) {
+    await deleteMerchantByUsername({ username: merchant.username })
   }
 }
